fix : view 영역을 침범하지 않고 카드 더 뽑도록 수정
