+++ src/main/kotlin/blackjack/domain/Drawable.kt
@@ -1,5 +1,5 @@
 package blackjack.domain
 
 interface Drawable {
-    fun canDraw(): Boolean
+    fun canDraw(player: Player): Boolean
 }
