+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -3,16 +3,15 @@ package blackjack.domain
 class Player(
     playerName: String,
     val battingAmount: Int,
-    playerCards: PlayerCards = PlayerCards()
+    playerCards: PlayerCards = PlayerCards(),
+    private val drawable: Drawable
 ) : Participant(name = playerName, participantCards = playerCards) {
-
-    private val canDrawable: Drawable = PlayerDrawable(this)
     override fun isDealer(): Boolean {
         return false
     }
 
     override fun canDrawable(): Boolean {
-        return participantCards.isHit() && canDrawable.canDraw()
+        return participantCards.isHit() && drawable.canDraw(this)
     }
 
     fun getEarnAmount(dealer: Dealer): Int {
