+++ src/test/kotlin/blackjack/domain/GameResultTest.kt
@@ -1,5 +1,6 @@
 package blackjack.domain
 
+import blackjack.view.GameView
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
@@ -7,7 +8,7 @@ class GameResultTest {
     @Test
     fun `참가자 배팅 10000, 딜러가 bust`() {
         val dealer = Dealer("딜러")
-        val players = Player("molly", 10000)
+        val players = Player("molly", 10000, drawable = GameView)
         BlackJackGame.of(
             dealer,
             listOf(players),
@@ -30,7 +31,7 @@ class GameResultTest {
     @Test
     fun `참가자 배팅 10000 플레이어가 bust`() {
         val dealer = Dealer("딜러")
-        val players = Player("molly", 10000)
+        val players = Player("molly", 10000, drawable = GameView)
         BlackJackGame.of(
             dealer,
             listOf(players),
@@ -53,7 +54,7 @@ class GameResultTest {
     @Test
     fun `플레이어가 10000원 배팅에 점수 12점, 딜러가 18점`() {
         val dealer = Dealer("딜러")
-        val players = Player("molly", 10000)
+        val players = Player("molly", 10000, drawable = GameView)
 
         val blackjack =
             BlackJackGame.of(dealer, listOf(players), MockCardDeck(Card(Card.CardPattern.CLUBS, Card.Denomination.SIX)))
@@ -67,7 +68,7 @@ class GameResultTest {
     fun `플레이어 3명 모두 동점일 때`() {
         val playerNames = listOf("molly", "jayce")
         val dealer = Dealer("딜러")
-        val players = playerNames.map { Player(it, 10000) }
+        val players = playerNames.map { Player(it, 10000, drawable = GameView) }
 
         val blackJackGame =
             BlackJackGame.of(dealer, players, MockCardDeck(Card(Card.CardPattern.CLUBS, Card.Denomination.EIGHT)))
@@ -80,7 +81,7 @@ class GameResultTest {
     @Test
     fun `플레이어가 블랙잭일때`() {
         val dealer = Dealer("딜러")
-        val player = Player("molly", 10000)
+        val player = Player("molly", 10000, drawable = GameView)
 
         val blackJackGame =
             BlackJackGame.of(
@@ -105,7 +106,7 @@ class GameResultTest {
     @Test
     fun `플레이어, 딜러 모두 블랙잭일때`() {
         val dealer = Dealer("딜러")
-        val player = Player("molly", 10000)
+        val player = Player("molly", 10000, drawable = GameView)
 
         val blackJackGame =
             BlackJackGame.of(
