+++ src/test/kotlin/blackjack/domain/PlayerTest.kt
@@ -1,5 +1,6 @@
 package blackjack.domain
 
+import blackjack.view.GameView
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
@@ -10,28 +11,28 @@ class PlayerTest {
     @ValueSource(strings = ["molly,jayce"])
     fun `참가자는 이름을 가진다`(input: String) {
         val players = input.split(",")
-        val participants = players.map { Player(it, 10000) }
+        val participants = players.map { Player(it, 10000, drawable = GameView) }
         assertThat(participants.size).isEqualTo(2)
     }
 
     @ParameterizedTest
     @ValueSource(strings = ["molly"])
     fun `참가자 최초 생성 시 스코어는 0점`(input: String) {
-        val player = Player(input, 10000)
+        val player = Player(input, 10000, drawable = GameView)
         assertThat(player.participantCards.score()).isEqualTo(0)
     }
 
     @ParameterizedTest
     @ValueSource(strings = ["molly"])
     fun `참가자 생성 직구 카드 갯수 0개`(input: String) {
-        val player = Player(input, 10000)
+        val player = Player(input, 10000, drawable = GameView)
         assertThat(player.participantCards.playerCards.size).isEqualTo(0)
     }
 
     @ParameterizedTest
     @ValueSource(strings = ["molly"])
     fun `참가자가 배팅한 금액 10000원`(input: String) {
-        val player = Player(input, 10000)
+        val player = Player(input, 10000, drawable = GameView)
         assertThat(player.battingAmount).isEqualTo(10000)
     }
 }
